[폴더위계정리] refactor: ext 와 util 마저 정리
