+++ app/src/main/java/com/wafflestudio/snutt2/data/mapper/NotificationMappers.kt
@@ -2,8 +2,8 @@ package com.wafflestudio.snutt2.data.mapper
 
 import com.wafflestudio.snutt2.domain.model.Notification
 import com.wafflestudio.snutt2.domain.model.NotificationType
+import com.wafflestudio.snutt2.lib.getLocalDateTimeFromString
 import com.wafflestudio.snutt2.network.dto.NotificationDto
-import com.wafflestudio.snutt2.ui.util.getLocalDateTimeFromString
 
 fun NotificationDto.toDomain(): Notification = Notification(
     title = title,
