+++ app/src/main/java/com/wafflestudio/snutt2/feature/friend/FriendsPage2.kt
@@ -61,7 +61,7 @@ import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
 import com.wafflestudio.snutt2.ui.theme.isDarkMode
-import com.wafflestudio.snutt2.ui.util.toFormattedString
+import com.wafflestudio.snutt2.ui.util.formatter.toFormattedString
 import com.wafflestudio.snutt2.ui.util.toast
 import kotlinx.coroutines.launch
