+++ app/src/main/java/com/wafflestudio/snutt2/feature/home/drawer/HomeDrawerContent.kt
@@ -35,7 +35,7 @@ import com.wafflestudio.snutt2.ui.components.compose.RedDot
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.toFormattedString
+import com.wafflestudio.snutt2.ui.util.formatter.toFormattedString
 
 /**
  * 논의
