+++ app/src/main/java/com/wafflestudio/snutt2/feature/home/drawer/bottom_sheet/CreateNewTableSheetContent.kt
@@ -34,7 +34,7 @@ import com.wafflestudio.snutt2.ui.components.compose.Picker
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.toFormattedString
+import com.wafflestudio.snutt2.ui.util.formatter.toFormattedString
 
 @Composable
 fun CreateTableBottomSheet(
