+++ app/src/main/java/com/wafflestudio/snutt2/feature/home/timetable/DrawClickEventDetector.kt
@@ -14,7 +14,7 @@ import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.input.pointer.pointerInteropFilter
 import com.wafflestudio.snutt2.domain.model.LocalLecture
 import com.wafflestudio.snutt2.domain.model.TableTrimParam
-import com.wafflestudio.snutt2.ui.util.contains
+import com.wafflestudio.snutt2.domain.model.contains
 
 @OptIn(ExperimentalComposeUiApi::class)
 @Composable
