+++ app/src/main/java/com/wafflestudio/snutt2/feature/login/SignUpPage.kt
@@ -43,14 +43,14 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.wafflestudio.snutt2.R
+import com.wafflestudio.snutt2.lib.isIdInvalid
+import com.wafflestudio.snutt2.lib.isPasswordInvalid
 import com.wafflestudio.snutt2.ui.components.compose.EditText
 import com.wafflestudio.snutt2.ui.components.compose.SimpleTopBar
 import com.wafflestudio.snutt2.ui.components.compose.WebViewStyleButton
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.isIdInvalid
-import com.wafflestudio.snutt2.ui.util.isPasswordInvalid
 import com.wafflestudio.snutt2.ui.util.toast
 
 @Composable
