+++ app/src/main/java/com/wafflestudio/snutt2/feature/search/LectureListItem.kt
@@ -42,9 +42,9 @@ import com.wafflestudio.snutt2.ui.components.compose.ThickReviewIcon
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.getLectureTagText
-import com.wafflestudio.snutt2.ui.util.getSimplifiedClassTimeForLecture
-import com.wafflestudio.snutt2.ui.util.getSimplifiedLocation
+import com.wafflestudio.snutt2.ui.util.formatter.getLectureTagText
+import com.wafflestudio.snutt2.ui.util.formatter.getSimplifiedClassTimeForLecture
+import com.wafflestudio.snutt2.ui.util.formatter.getSimplifiedLocation
 
 @Composable
 fun ExpandableLectureListItem(
