+++ app/src/main/java/com/wafflestudio/snutt2/feature/search/SearchTagCell.kt
@@ -25,7 +25,7 @@ import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.components.compose.displayName
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.color
+import com.wafflestudio.snutt2.ui.util.formatter.color
 
 @Composable
 fun SearchTagCell(
