+++ app/src/main/java/com/wafflestudio/snutt2/feature/search/search_option/TimeSelectSheet.kt
@@ -47,6 +47,7 @@ import com.wafflestudio.snutt2.domain.model.LocalLecture
 import com.wafflestudio.snutt2.domain.model.TableLectureCustom
 import com.wafflestudio.snutt2.domain.model.TableTrimParam
 import com.wafflestudio.snutt2.domain.model.preview.PreviewData
+import com.wafflestudio.snutt2.domain.model.trimByTrimParam
 import com.wafflestudio.snutt2.feature.home.timetable.DrawClassTime
 import com.wafflestudio.snutt2.feature.home.timetable.DrawTableGrid
 import com.wafflestudio.snutt2.feature.home.timetable.TimetableCanvasObjects
@@ -57,7 +58,6 @@ import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
 import com.wafflestudio.snutt2.ui.theme.isDarkMode
 import com.wafflestudio.snutt2.ui.theme.onSurfaceVariant
-import com.wafflestudio.snutt2.ui.util.trimByTrimParam
 
 @Composable
 fun TimeSelectSheet(
