+++ app/src/main/java/com/wafflestudio/snutt2/feature/settings/UserConfigViewModel.kt
@@ -8,8 +8,8 @@ import com.wafflestudio.snutt2.data.user.UserRepository
 import com.wafflestudio.snutt2.domain.AuthError
 import com.wafflestudio.snutt2.domain.DisplayMessageResolver
 import com.wafflestudio.snutt2.domain.DomainError
-import com.wafflestudio.snutt2.ui.util.isIdInvalid
-import com.wafflestudio.snutt2.ui.util.isPasswordInvalid
+import com.wafflestudio.snutt2.lib.isIdInvalid
+import com.wafflestudio.snutt2.lib.isPasswordInvalid
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.MutableStateFlow
