+++ app/src/main/java/com/wafflestudio/snutt2/feature/table_lectures/TableLecturesScreen.kt
@@ -39,10 +39,10 @@ import com.wafflestudio.snutt2.ui.components.compose.TagIcon
 import com.wafflestudio.snutt2.ui.components.compose.clicks
 import com.wafflestudio.snutt2.ui.theme.SNUTTColors
 import com.wafflestudio.snutt2.ui.theme.SNUTTTypography
-import com.wafflestudio.snutt2.ui.util.getInstructorAndCreditText
-import com.wafflestudio.snutt2.ui.util.getLectureTagText
-import com.wafflestudio.snutt2.ui.util.getSimplifiedClassTimeForLecture
-import com.wafflestudio.snutt2.ui.util.getSimplifiedLocation
+import com.wafflestudio.snutt2.ui.util.formatter.getInstructorAndCreditText
+import com.wafflestudio.snutt2.ui.util.formatter.getLectureTagText
+import com.wafflestudio.snutt2.ui.util.formatter.getSimplifiedClassTimeForLecture
+import com.wafflestudio.snutt2.ui.util.formatter.getSimplifiedLocation
 
 @Composable
 fun TableLecturesRoute(
